perf(ctx): report_progress uses uint8_t 0-100 instead of double — eliminates all floating-point from progress path
- ExecutionContext::report_progress takes uint8_t pct instead of double
- _progress: atomic<double> → atomic<uint8_t> (cheaper loads/stores)
- _progress_pct: atomic<int> → atomic<int8_t>
- ProgressPayload::pct: double → uint8_t
- AlgorithmObserver::on_progress takes uint8_t pct
- All strategy callers updated: 0.0→0, 1.0→100, dynamic formulas converted to integer math
- progress() returns double via pct/100.0 conversion

+++ tests/integration/test_algorithm_executor.cpp
@@ -21,7 +21,7 @@ public:
         for (int i = 0; i < 5; i++) {
             if (ctx.is_cancelled()) return;
             ctx.wait_if_paused();
-            ctx.report_progress((i + 1) * 20.0, ProgressStatus::Exploring);
+            ctx.report_progress(static_cast<uint8_t>((i + 1) * 20), ProgressStatus::Exploring);
             std::this_thread::sleep_for(std::chrono::milliseconds(20));
         }
         std::vector<compact::EnchStep> solution;
